add async to locationservice and location controller

+++ wms-praktyki-yosi-api/wms-praktyki-yosi-api/Controllers/LocationController.cs
@@ -14,13 +14,19 @@ namespace wms_praktyki_yosi_api.Controllers
     {
 
         private readonly ILocationService _locationService;
+        private readonly ICustomAuthorizationService _authorizationService;
 
-        public LocationController(ILocationService locationService)
+        public LocationController(ILocationService locationService, ICustomAuthorizationService authorizationService)
         {
             _locationService = locationService;
+            _authorizationService = authorizationService;
+
         }
 
-        [HttpGet] public ActionResult GetAll() {
+        [HttpGet] public async Task<ActionResult> GetAll() {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             var result = _locationService.GetAllLocations();
             if (result == null)
             {
@@ -30,12 +36,14 @@ namespace wms_praktyki_yosi_api.Controllers
         }
         [HttpPost]
         [Authorize(Roles = "Admin,Moderator")]
-        public ActionResult AddProductToLocation([FromBody] ProductLocationDto dto)
+        public async Task<ActionResult> AddProductToLocation([FromBody] ProductLocationDto dto)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
             try
             {
                 var result = _locationService.AddProductToLocation(dto);
-                return Ok();
+                return Ok(result);
             }
             catch(BadRequestException ex)
             {
@@ -45,14 +53,19 @@ namespace wms_praktyki_yosi_api.Controllers
           
         }
         [HttpGet("{id}")]
-        [Authorize(Roles = "Admin,Moderator")] public ActionResult GetById([FromRoute] int id)
+        [Authorize(Roles = "Admin,Moderator")] public async Task<ActionResult> GetById([FromRoute] int id)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
             var result = _locationService.GetLocationById(id);
             return Ok(result);
         }
         [HttpPut("{id}")]
         [Authorize(Roles = "Admin,Moderator")]
-        public ActionResult Edit([FromRoute] int id, [FromBody] ProductLocationDto dto) {
+        public async Task<ActionResult> Edit([FromRoute] int id, [FromBody] ProductLocationDto dto) {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             var result = _locationService.UpdateLocation(id, dto);
             if (!result)
             {
@@ -62,7 +75,10 @@ namespace wms_praktyki_yosi_api.Controllers
         }
         [HttpDelete("{id}")]
         [Authorize(Roles = "Admin,Moderator")]
-        public ActionResult Delete([FromRoute] int id) {
+        public async Task<ActionResult> Delete([FromRoute] int id) {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             var result = _locationService.DeleteLocation(id);
             if (!result)
             {
